apache2.initd: capture config test output in checkconfig().
The configtest() command in the init script calls "apache2 -t" to test
the configuration for syntax errors. Before this commit, the command
was run twice -- once to see if there was an error, and once to output
the error to the user. We can capture the output in a variable the
first time; this commit does so, and prints the contents of that
variable when there is an error.