apache2.initd: remove config file readability check.
In checkconfd(), the apache2 init script checks to make sure that
$CONFIGFILE is readable. To do this, it needs to reimplement the
relative path handling that apache2 already does. If we remove the
relative path handling from checkconfd(), then apache2 continues to do
the right thing when $CONFIGFILE is relative. However, our custom
readability check fails, since it is looking for an absolute path
that doesn't exist.

Fortunately, apache2 already checks to ensure that the configuration
file exists and is readable. If we remove the custom readability check
entirely, then we get an error like,

   * apache2 has detected an error in your setup:
  apache2: Could not open configuration file
  /usr/lib/apache2/example.conf: No such file or directory

This is not sufficiently worse than the custom error:

  Unable to read configuration file: example.conf

Therefore the custom readability check is removed in this commit, and
we can avoid reimplementing the relative path handling as well.