apache2.initd: append STARTUPERRORLOG to APACHE2_OPTS in start().
We currently append the value of STARTUPERRORLOG to APACHE2_OPTS in
the checkconfd() function. This can cause duplication of command-line
arguments as witnessed in bug 566726. Since STARTUPERRORLOG does need
conditional handling (we don't append it if it's empty), the wisest
place to do so seems to be in start().

Gentoo-Bug: 566726