apache2.initd: append SERVERROOT and CONFIGFILE to APACHE2_OPTS once.
We are currently appending both SERVERROOT and CONFIGFILE to
APACHE2_OPTS in the checkconfd() function. Since checkconfd() is
called multiple times, this can lead to duplication of command-line
arguments as witnessed in bug 566726. Since the values of SERVERROOT
and CONFIGFILE are fixed, we now instead append them once at the
beginning of the script.

Gentoo-Bug: 566726