apache2.initd: set defaults for LYNX and STATUSURL once.
The default values for the LYNX and STATUSURL variables were set in
functions. Now that the other conf.d defaults are set at the beginning
of the script, it makes sense to set these there, too. The LYNX
default was set twice, in configdump() and fullstatus(), so this
eliminates some duplication.